Refactor for optional dependencies
