+++ pyproject.toml
@@ -28,6 +28,10 @@ classifiers = [
 "Bug Tracker" = "https://github.com/kuznia-rdzeni/transactron/issues"
 
 [project.optional-dependencies]
+testing = [
+    "pytest >= 8.0.0, < 10.0",
+    "hypothesis <= 6.99.6, < 6.152.6",
+]
 dev = [
     # Linting
     "pyright == 1.1.393",
