+++ test/lib/test_stack.py
@@ -1,7 +1,8 @@
 import pytest
 import random
 from transactron.lib.stack import Stack
-from transactron.testing import TestCaseWithSimulator, data_layout, TestbenchContext, SimpleTestCircuit
+from transactron.testing import data_layout, TestbenchContext, SimpleTestCircuit
+from transactron.testing.test_case import TestCaseWithSimulator
 
 
 class TestStack(TestCaseWithSimulator):
