+++ test/testing/test_log.py
@@ -10,7 +10,8 @@ from amaranth import *
 from amaranth.lib.enum import Enum, EnumView
 
 from transactron import *
-from transactron.testing import TestCaseWithSimulator, TestbenchContext
+from transactron.testing import TestbenchContext
+from transactron.testing.test_case import TestCaseWithSimulator
 from transactron.utils import logging
 from transactron.testing.logging import HDLLogWrapper, make_logging_process
