+++ test/testing/test_method_mock.py
@@ -5,8 +5,9 @@ from amaranth.lib.data import StructLayout
 import pytest
 
 from transactron import *
-from transactron.testing import TestCaseWithSimulator, TestbenchContext, SimpleTestCircuit
+from transactron.testing import TestbenchContext, SimpleTestCircuit
 from transactron.testing.method_mock import MethodMock, def_method_mock
+from transactron.testing.test_case import TestCaseWithSimulator
 from transactron.lib import *
