+++ test/testing/test_validate_arguments.py
@@ -2,7 +2,8 @@ import random
 from amaranth import *
 from amaranth.sim import *
 
-from transactron.testing import TestCaseWithSimulator, TestbenchIO, data_layout, TestbenchContext
+from transactron.testing import TestbenchIO, data_layout, TestbenchContext
+from transactron.testing.test_case import TestCaseWithSimulator
 
 from transactron import *
 from transactron.testing.method_mock import def_method_mock
