+++ test/utils/amaranth_ext/test_elaboratables.py
@@ -2,7 +2,8 @@ import pytest
 import random
 
 from transactron.utils import OneHotRoundRobin, StableSelectingNetwork
-from transactron.testing import TestCaseWithSimulator, TestbenchContext
+from transactron.testing import TestbenchContext
+from transactron.testing.test_case import TestCaseWithSimulator
 
 
 class TestScheduler(TestCaseWithSimulator):
