+++ transactron/testing/__init__.py
@@ -1,6 +1,7 @@
 from amaranth.sim._async import TestbenchContext, ProcessContext, SimulatorContext  # noqa: F401
-from .input_generation import *  # noqa: F401
-from .infrastructure import *  # noqa: F401
+from transactron.utils import data_layout  # noqa: F401
+
+# .input_generation and .infrastructure are not reexported because of optional dependencies
 from .functions import *  # noqa: F401
 from .simulator import *  # noqa: F401
 from .test_circuit import *  # noqa: F401
@@ -9,4 +10,3 @@ from .testbenchio import *  # noqa: F401
 from .profiler import *  # noqa: F401
 from .logging import *  # noqa: F401
 from .tick_count import *  # noqa: F401
-from transactron.utils import data_layout  # noqa: F401
